requirements.txt
# -----------------------------------------------
# Core Framework (GPU build - CUDA 11.8)
# -----------------------------------------------
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.9.0

# -----------------------------------------------
# Hugging Face Ecosystem
# -----------------------------------------------
transformers>=4.56.1
datasets>=4.0.0
accelerate>=1.11.0
peft>=0.17.1
hf_transfer
hf_xet

# -----------------------------------------------
# Quantization and Memory-Efficient Training
# -----------------------------------------------
bitsandbytes>=0.42.0

# -----------------------------------------------
# Evaluation and Metrics
# -----------------------------------------------
evaluate>=0.4.0
rouge-score
lm-eval

# -----------------------------------------------
# Experiment Tracking and Configuration
# -----------------------------------------------
wandb>=0.22.0
python-dotenv>=1.1.1
PyYAML>=6.0.2

# -----------------------------------------------
# Utilities
# -----------------------------------------------
tqdm>=4.65.0
numpy>=1.26.0

# -----------------------------------------------
# API Integration
# -----------------------------------------------
openai>=2.7.0


